Membuat class Personal Information untuk mengatasi bad smell data clumps string pada file Employee Information dan dipindahkan ke personal information
